# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Jimce API",
    "version": "0.0.0-dev"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token sent as Authorization: Bearer <token>"
      }
    },
    "schemas": {
      "400-bad-request-response": {
        "description": "A response sent by the server when your request is malformed (Code 400)",
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "const": 400
          },
          "code": {
            "description": "Internal code describing the error",
            "examples": [
              "FST_ERR_VALIDATION"
            ],
            "type": "string"
          },
          "error": {
            "type": "string",
            "const": "Bad Request"
          },
          "message": {
            "description": "Error message describing what went wrong",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "error",
          "message"
        ],
        "additionalProperties": false
      },
      "401-unauthorized-response": {
        "description": "A response sent by the server when your request is not authorized",
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "const": 401
          },
          "code": {
            "description": "Internal code describing the error",
            "examples": [
              "FST_JWT_NO_AUTHORIZATION_IN_HEADER"
            ],
            "type": "string"
          },
          "error": {
            "type": "string",
            "const": "Unauthorized"
          },
          "message": {
            "description": "Error message describing what went wrong",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "error",
          "message"
        ],
        "additionalProperties": false
      },
      "403-forbidden-response": {
        "description": "A response sent by the server when lack the required permissions (HTTP 403)",
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "const": 403
          },
          "code": {
            "description": "Internal code describing the error",
            "examples": [
              "NOT_AN_ADMIN"
            ],
            "type": "string"
          },
          "error": {
            "type": "string",
            "const": "Forbidden"
          },
          "message": {
            "description": "Error message describing what went wrong",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "error",
          "message"
        ],
        "additionalProperties": false
      },
      "500-internal-server-error-response": {
        "description": "A response sent by the server when the server failed to process your request",
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "const": 500
          },
          "code": {
            "description": "Internal code describing the error",
            "examples": [
              "FST_ERR_RESPONSE_SERIALIZATION"
            ],
            "type": "string"
          },
          "error": {
            "type": "string",
            "const": "Internal Server Error"
          },
          "message": {
            "description": "Error message describing what went wrong",
            "examples": [
              "Response does not match the schema"
            ],
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "error",
          "message"
        ],
        "additionalProperties": false
      },
      "success-indicator-response": {
        "description": "A response sent by the server when your request is valid to indicate whether it was successful or not",
        "type": "object",
        "properties": {
          "success": {
            "description": "States whether the requested action was successful",
            "type": "boolean"
          },
          "reason": {
            "description": "An optional reason for the success or fail of your requested action",
            "examples": [
              "User already exists"
            ],
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
    "/api/admin/users/list-users": {
      "get": {
        "summary": "Lists all users",
        "description": "Returns id, username, email and isAdmin for every registered user\n                \n                **Requires authentication**\n                **Requires admin privileges** \n                ",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "description": "The uuid of the user",
                        "examples": [
                          "5b9e340c-38ba-4495-b895-c3aee230134e"
                        ],
                        "type": "string"
                      },
                      "username": {
                        "type": "string"
                      },
                      "email": {
                        "description": "If existent, the email of the user",
                        "examples": [
                          null,
                          "user@example.com"
                        ],
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "isAdmin": {
                        "description": "Defines whether the user has admin privileges",
                        "examples": [
                          false
                        ],
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "username",
                      "email",
                      "isAdmin"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/400-bad-request-response"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/401-unauthorized-response"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/403-forbidden-response"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/500-internal-server-error-response"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/create-or-change": {
      "put": {
        "summary": "Creates or changes a new user",
        "description": "Based on provided username, password, email and isAdmin bool, this route will create or update a new user.\n                When the user with the provided username already exists, the user will be updated instead of re-created.\n            \n                **Requires authentication**\n                **Requires admin privileges** \n            ",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "description": "The username which will be assigned to the user (if already in use, user will be updated)",
                    "examples": [
                      "test.user"
                    ],
                    "type": "string",
                    "minLength": 3
                  },
                  "password": {
                    "description": "The new password in plain text for the user. Optional when user already exists, necessary when creating from scratch. Must be at least 12 characters long",
                    "type": "string",
                    "minLength": 12
                  },
                  "email": {
                    "description": "If provided, the new email of the user, if set to null, it will still be updated to null.",
                    "examples": [
                      "user@example.com"
                    ],
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "email",
                        "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "isAdmin": {
                    "description": "Declares whether the user shall have admin rights",
                    "examples": [
                      false,
                      true
                    ],
                    "type": "boolean"
                  }
                },
                "required": [
                  "username",
                  "email",
                  "isAdmin"
                ]
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success-indicator-response"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/400-bad-request-response"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/401-unauthorized-response"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/403-forbidden-response"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/500-internal-server-error-response"
                }
              }
            }
          }
        }
      }
    },
...
  }
}